# backend/zetipay_backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ZetiPay REST API for wallet connection, crypto transfers, and asset management.",
        "title": "ZetiPay API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5000",
    "basePath": "/api",
    "paths": {
        "/assets/{address}": {
            "get": {
                "description": "Retrieves ETH and token balances, transaction history",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets"
                ],
                "summary": "Get wallet assets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wallet address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.AssetsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/gas/estimate": {
            "post": {
                "description": "Estimates gas for ETH or token transfer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets"
                ],
                "summary": "Estimate gas for a transaction",
                "parameters": [
                    {
                        "description": "Transaction details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.GasEstimateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.GasEstimateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/transfer/eth": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Initiates an Ethereum transfer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transfer"
                ],
                "summary": "Transfer ETH",
                "parameters": [
                    {
                        "description": "ETH transfer details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.TransferETHRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.TransferResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/transfer/token": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Initiates an ERC-20 token transfer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transfer"
                ],
                "summary": "Transfer ERC-20 token",
                "parameters": [
                    {
                        "description": "Token transfer details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.TransferTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.TransferResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/wallet/connect": {
            "post": {
                "description": "Connects an Ethereum wallet and retrieves its balances",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "Connect a wallet",
                "parameters": [
                    {
                        "description": "Wallet address",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.WalletConnectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.WalletConnectResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.AssetsResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "dbTransactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Transaction"
                    }
                },
                "ethBalance": {
                    "type": "string"
                },
                "tokenBalances": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "handlers.GasEstimateRequest": {
            "type": "object",
            "required": [
                "amount",
                "fromAddress",
                "toAddress"
            ],
            "properties": {
                "amount": {
                    "type": "string"
                },
                "fromAddress": {
                    "type": "string"
                },
                "toAddress": {
                    "type": "string"
                },
                "tokenAddress": {
                    "type": "string"
                }
            }
        },
        "handlers.GasEstimateResponse": {
            "type": "object",
            "properties": {
                "gasEstimate": {
                    "type": "string"
                }
            }
        },
        "handlers.TransferETHRequest": {
            "type": "object",
            "required": [
                "amount",
                "fromAddress",
                "signature",
                "toAddress"
            ],
            "properties": {
                "amount": {
                    "type": "string"
                },
                "fromAddress": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                },
                "toAddress": {
                    "type": "string"
                }
            }
        },
        "handlers.TransferResponse": {
            "type": "object",
            "properties": {
                "transactionHash": {
                    "type": "string"
                }
            }
        },
        "handlers.TransferTokenRequest": {
            "type": "object",
            "required": [
                "amount",
                "fromAddress",
                "signature",
                "toAddress",
                "tokenAddress"
            ],
            "properties": {
                "amount": {
                    "type": "string"
                },
                "fromAddress": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                },
                "toAddress": {
                    "type": "string"
                },
                "tokenAddress": {
                    "type": "string"
                }
            }
        },
        "handlers.WalletConnectRequest": {
            "type": "object",
            "required": [
                "address"
            ],
            "properties": {
                "address": {
                    "type": "string"
                }
            }
        },
        "handlers.WalletConnectResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "ethBalance": {
                    "type": "string"
                },
                "tokenBalances": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "models.Transaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "toAddress": {
                    "type": "string"
                },
                "tokenAddress": {
                    "type": "string"
                },
                "txHash": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "walletAddress": {
                    "type": "string"
                }
            }
        }
    }
}